% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaParams.r
\name{beaParams}
\alias{beaParams}
\title{Gives list of parameters possible for a given dataset}
\usage{
beaParams(beaKey, setName)
}
\arguments{
\item{beaKey}{Your API key}

\item{setName}{Name of BEA dataset (e.g., 'NIPA')}
}
\value{
A metadata object of class 'list' of several dimensions. View list structure using 'str(yourList)'.
}
\description{
Gives list of parameters possible for a given dataset
}
\examples{
beaParams('yourAPIkey', 'RegionalData')
}
\keyword{metadata}
