% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaViz.r
\name{beaViz}
\alias{beaViz}
\title{Visualize BEA API response payload}
\usage{
beaViz(beaPayload)
}
\arguments{
\item{beaPayload}{An httr response from call to BEA API}
}
\description{
CAUTION: Currently only works with NATIONAL datasets (NIPA, NIUnderlyingDetail, FixedAssets)
}
\examples{
userSpecList <- list('UserID' = 'yourKey' ,
								'Method' = 'GetData',
								'datasetname' = 'NIPA',
								'Frequency' = 'A',
								'TableID' = '68',
								'Year' = 'X')
resp <- beaGet(userSpecList)
BDF <- beaViz(resp)
userSpecList <- list('UserID' = 'yourKey' ,
								'Method' = 'GetData',
								'datasetname' = 'NIPA',
								'Frequency' = 'A',
								'TableID' = '68',
								'Year' = 'X')
}

