% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaSets.r
\name{beaSets}
\alias{beaSets}
\title{Returns a list of all datasets}
\usage{
beaSets(beaKey)
}
\arguments{
\item{beaKey}{Your API key}
}
\value{
A metadata object of class 'list' of several dimensions. View list structure using 'str(yourList)'.
}
\description{
Returns a list of all datasets
}
\examples{
beaSets('yourAPIkey')
}
\keyword{metadata}
