% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaUpdateMetadata.r
\name{beaUpdateMetadata}
\alias{beaUpdateMetadata}
\title{Download BEA metadata into library/data folder if needed}
\usage{
beaUpdateMetadata(datasetList, beaKey)
}
\arguments{
\item{datasetList}{list of BEA datasets to update local metadata file for (e.g., list('NIPA', 'FixedAssets'))}

\item{beaKey}{Your API key}
}
\value{
Nothing. This updates local .RData files to be used in beaSearch.
}
\description{
Download BEA metadata into library/data folder if needed
}
\examples{
beaUpdateMetadata(list('RegionalData', 'NIPA'), beaKey = 'yourAPIkey')
}
\keyword{metadata}
\keyword{search}
