% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaViz.r
\name{beaViz}
\alias{beaViz}
\title{Visualize BEA API response payload}
\usage{
beaViz(beaPayload = NULL, beaKey = NULL)
}
\arguments{
\item{beaPayload}{An httr response from call to BEA API}

\item{beaKey}{Your 36-digit BEA API key}
}
\description{
When entered into the R console, the function below starts an interactive dashboard. CAUTION: Currently only works with NATIONAL datasets (NIPA, NIUnderlyingDetail, FixedAs-sets). R Studio users must opt to "show in browser" for this method to be fully functional.
}
\examples{
userSpecList <- list('UserID' = 'yourKey' ,
								'Method' = 'GetData',
								'datasetname' = 'NIPA',
								'Frequency' = 'A',
								'TableID' = '68',
								'Year' = 'X')		
resp <- beaGet(userSpecList)
BDF <- beaViz(resp)
userSpecList <- list('UserID' = 'yourKey' ,
								'Method' = 'GetData',
								'datasetname' = 'NIPA',
								'Frequency' = 'A',
								'TableID' = '68',
								'Year' = 'X')
}
